Finish remaining long tail test files
